requirements.txt
absl-py==0.2.0
astor==0.6.2
bleach==1.5.0
Django==2.0.4
django-cors-headers==2.2.0
djangorestframework==3.8.2
gast==0.2.0
grpcio==1.11.0
html5lib==0.9999999
Markdown==2.6.11
numpy==1.14.3
pkg-resources==0.0.0
protobuf==3.5.2.post1
pytz==2018.4
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
Werkzeug==0.14.1
